server/requirements/base.txt
django==4.0.2
django-environ==0.8.1
django-allauth==0.49.0
django-taggit==3.0.0
argon2-cffi==21.3.0
Pillow==9.0.1

djangorestframework==3.13.1
dj-rest-auth==2.2.3
djangorestframework-simplejwt==5.0.0
django-filter==21.1
drf-extra-fields==3.2.1
